+++ AccountManagment.Domain/RoleAgg/Permission.cs
@@ -2,10 +2,10 @@
 {
     public class Permission
     {
-        public long Id { get; private set; }
+        public int Id { get; private set; }
         public int Code { get; private set; }
         public string Name { get; private set; }
-        public long RoleId { get; private set; }
+        public int RoleId { get; private set; }
         public Role Role { get; private set; }
 
         public Permission(int code)
